+++ x/crosschain/types/status.go
@@ -5,24 +5,25 @@ import (
 	"slices"
 )
 
+type StatusMessages struct {
+	StatusMessage        string `json:"status_message"`
+	ErrorMessageOutbound string `json:"error_message_outbound"`
+	ErrorMessageRevert   string `json:"error_message_revert"`
+}
+
 func (m *Status) AbortRefunded() {
 	m.IsAbortRefunded = true
 	m.StatusMessage = "CCTX aborted and Refunded"
 }
 
-// UpdateStatusAndErrorMessages transitions the Status and Error messages.
-func (m *Status) UpdateStatusAndErrorMessages(newStatus CctxStatus, statusMsg, errorMsg string) {
-	m.UpdateStatus(newStatus, statusMsg)
-
-	if errorMsg != "" {
-		m.UpdateErrorMessage(errorMsg)
-	}
+func (m *Status) UpdateStatusAndErrorMessages(newStatus CctxStatus, messages StatusMessages) {
+	m.UpdateStatus(newStatus)
+	m.UpdateErrorMessages(messages)
 }
 
-// UpdateStatus updates the cctx status and cctx.status.status_message.
-func (m *Status) UpdateStatus(newStatus CctxStatus, statusMsg string) {
+// UpdateStatus updates the cctx status
+func (m *Status) UpdateStatus(newStatus CctxStatus) {
 	if m.ValidateTransition(newStatus) {
-		m.StatusMessage = fmt.Sprintf("Status changed from %s to %s", m.Status.String(), newStatus.String())
 		m.Status = newStatus
 	} else {
 		m.StatusMessage = fmt.Sprintf(
@@ -30,24 +31,21 @@ func (m *Status) UpdateStatus(newStatus CctxStatus, statusMsg string) {
 			m.Status.String(),
 			newStatus.String(),
 		)
-
 		m.Status = CctxStatus_Aborted
-	}
-
-	if statusMsg != "" {
-		m.StatusMessage += fmt.Sprintf(": %s", statusMsg)
 	}
 }
 
-// UpdateErrorMessage updates cctx.status.error_message.
-func (m *Status) UpdateErrorMessage(errorMsg string) {
-	errMsg := errorMsg
+// UpdateErrorMessages updates cctx.status.error_message and cctx.status.error_message_revert.
+func (m *Status) UpdateErrorMessages(messages StatusMessages) {
+	// Always update the status message, status should contain only the most recent update
+	m.StatusMessage = messages.StatusMessage
 
-	if errMsg == "" {
-		errMsg = "unknown error"
+	if messages.ErrorMessageOutbound != "" {
+		m.ErrorMessage = messages.ErrorMessageOutbound
 	}
-
-	m.ErrorMessage = errMsg
+	if messages.ErrorMessageRevert != "" {
+		m.ErrorMessageRevert = messages.ErrorMessageRevert
+	}
 }
 
 func (m *Status) ValidateTransition(newStatus CctxStatus) bool {